requirements.txt
pyyaml>=6.0.1
python-dotenv>=1.0.1
openai>=1.48.0
numpy>=1.26.0
pypdf>=4.2.0
python-docx>=1.1.2
rank-bm25>=0.2.2
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

